+++ modules/patch-management/r-patchmanagement.tf
@@ -13,10 +13,11 @@ resource "azurerm_log_analytics_solution" "patch_management" {
   }
 }
 
-resource "azurerm_template_deployment" "update_config_standard_patch_linux" {
-  count               = contains([for v in var.patch_mgmt_os : lower(v)], "linux") ? 1 : 0
+resource "azurerm_template_deployment" "update_config_standard_patch" {
+  for_each = toset(var.patch_mgmt_os)
+
   deployment_mode     = "Incremental"
-  name                = format("%s-%s", local.arm_patch_mngt_name, "linux")
+  name                = lower(format("%s-%s", local.arm_patch_mngt_name, each.key))
   resource_group_name = var.resource_group_name
 
   template_body = jsonencode({
@@ -26,70 +27,28 @@ resource "azurerm_template_deployment" "update_config_standard_patch_linux" {
       {
         type       = "Microsoft.Automation/automationAccounts/softwareUpdateConfigurations"
         apiVersion = "2019-06-01"
-        name       = "${var.automation_account_name}/Standard Linux Update Schedule"
+        name       = "${var.automation_account_name}/Standard ${title(each.key)} Update Schedule"
         properties = {
           updateConfiguration = {
-            operatingSystem = "Linux"
+            operatingSystem = "${each.key}"
             linux = {
-              includedPackageClassifications : "${join(", ", var.patch_mgmt_linux_update_classifications)}"
-              rebootSetting = "${var.patch_mgmt_linux_reboot_setting}"
+              includedPackageClassifications : "${join(", ", var.patch_mgmt_update_classifications)}"
+              rebootSetting = "${var.patch_mgmt_reboot_setting}"
             }
-            duration = "${var.patch_mgmt_linux_duration}"
+            duration = "${var.patch_mgmt_duration}"
             targets = {
               azureQueries = [
                 {
-                  scope = "${var.patch_mgmt_linux_scope}"
+                  scope = "${var.patch_mgmt_scope}"
                   tagSettings = {
-                    tags           = "${var.patch_mgmt_linux_tags_filtering}"
-                    filterOperator = "${var.patch_mgmt_linux_tags_filtering_operator}"
+                    tags           = "${var.patch_mgmt_tags_filtering}"
+                    filterOperator = "${var.patch_mgmt_tags_filtering_operator}"
                   }
                 }
               ]
             }
           }
-          scheduleInfo = "${var.patch_mgmt_linux_schedule[0]}"
-        }
-      }
-    ]
-  })
-}
-
-
-resource "azurerm_template_deployment" "update_config_standard_patch_windows" {
-  count               = contains([for v in var.patch_mgmt_os : lower(v)], "windows") ? 1 : 0
-  deployment_mode     = "Incremental"
-  name                = format("%s-%s", local.arm_patch_mngt_name, "windows")
-  resource_group_name = var.resource_group_name
-
-  template_body = jsonencode({
-    "$schema"      = "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#"
-    contentVersion = "1.0.0.0"
-    resources = [
-      {
-        type       = "Microsoft.Automation/automationAccounts/softwareUpdateConfigurations"
-        apiVersion = "2019-06-01"
-        name       = "${var.automation_account_name}/Standard Windows Update Schedule"
-        properties = {
-          updateConfiguration = {
-            operatingSystem = "Windows"
-            linux = {
-              includedPackageClassifications : "${join(", ", var.patch_mgmt_windows_update_classifications)}"
-              rebootSetting = "${var.patch_mgmt_windows_reboot_setting}"
-            }
-            duration = "${var.patch_mgmt_windows_duration}"
-            targets = {
-              azureQueries = [
-                {
-                  scope = "${var.patch_mgmt_windows_scope}"
-                  tagSettings = {
-                    tags           = "${var.patch_mgmt_windows_tags_filtering}"
-                    filterOperator = "${var.patch_mgmt_windows_tags_filtering_operator}"
-                  }
-                }
-              ]
-            }
-          }
-          scheduleInfo = "${var.patch_mgmt_windows_schedule[0]}"
+          scheduleInfo = "${var.patch_mgmt_schedule[0]}"
         }
       }
     ]